Add processor for evalai accuracy
